Modify regex pattern that finds sentence boundaries.
The old  regex pattern does not account for a period,
exclamation point, or question mark that is immediately followed by
a non-alphabetical character (like quotation marks).

The string:

"and what is the use of a book," thought Alice
"without pictures or conversation?"

will yield two sentences, instead of one because
the question mark splits the string into the desired sentence,
and the exclamation point at the end.

The new regex pattern eliminates this undesired behaviour.