Change data structure for candidate objects from list to a dictionary.
Previous structure is a list of candidate objects, i.e.
[Candidate1, Candidate2, ...]

New Structure is a dictionary whose keys are the names of candidates,
and whose values are the candidate object itself
(with attributes: name, votes, elimination status)
i.e.
{"name1": Candidate1, "name2": Candidate2, ...}

This is done to make checking for a valid candidate easier.
Instead of iterating through the list, then checking the name,
The keys() function can be called and membership checked.